Add Option for Dropping Leading H1 Heading (#36)
* Add goland IDE prefs to .gitignore

* Add flag for removing leading H1 heading in markdown; fixes #35

* Remove manpage reference to unused -n and -c args

* Add manpage docs for --drop-h1 argument

* Update docs; add docs for --drop-h1 argument

* Rename method to be more verbose; add code comments

Renames to `DropDocumentLeadingH1` to better denote the method's
purpose and adds code docs to this effect in order to hopefully
negate potential misuse on single lines.

* Revert "Add manpage docs for --drop-h1 argument"

This reverts commit 48a641baed05540f6172fe9b780b06eacad4dc28.

The `--drop-h1` flag is not understood as an argument in this
context, but rather is folded under `[options]`. This change
caused the arguments to not be properly parsed when the --drop-h1
flag was passed.