fix: Remove double interpolation for attachments (BREAKING)
The use case was that when using the "attachment://", it would replace
the link twice, once on the "attachment://" and after that on the
replace all as well as the string will be present in the replacement.
This ensures that the keyword only get replcaed once.

Furthermore, put "attachment://" in as a identifier for a replacement is
better then nothing since you effectively create a ban word in the name
of the attachement file.